Add files via upload
This is still very incomplete, as it is just a modification of a database I made for an internship last year. Lines 15-84 are what's actually relevant to our project. The rest I'm just gonna remove or modify to fit our project's specifications.
Haven't messed around with stored procedures or inserting dummy data yet, will do when I have a better idea of what the database is supposed to look like.